✅ [DONE] Product crud api

+++ controllers/categoryController.js
@@ -66,9 +66,15 @@ exports.updateCategory = async (req, res) => {
 }
 
 exports.deleteCategory = async (req, res) => {
+
     try {
         if (!await prisma.category.findUnique({ where: { id: parseInt(req.params.id) } })) {
             return res.status(404).json({ error: 'Category not found' })
+        }
+
+        const productCount = await prisma.product.count({ where: { categoryId: parseInt(req.params.id) } })
+        if (productCount) {
+            return res.status(409).json({ error: `Category id being used in ${productCount} product(s)` })
         }
 
         await prisma.category.delete({ where: { id: parseInt(req.params.id) } })
